Temporary solution to #2.
The real issue is that the parser use position argument with narg="*"
to parse any non-listed git commands. But the parser would stop parsing
position args when is sees a flag "-x", "--flag". Since this argument is
not recorded via add_argument. It says it's an unrecognized args.

The current solution just adds a great amount of git built-in commands
to the list of EchoParser, which does no real parsing but just returns
the tokens. However, this will not work in general for customized git
command as they are again not handled by EchoParser.